Remove `os.Exec()` calls from `check-logs.go`
If we are to be using Golang for some of the work in the log
functionality test, we should make sure to delegate only the
work that is not possible or not simple in Bash. The calls to
`os.Exec()` are perhaps more convoluted in the Golang program
than they are in Bash, so this change separates them out into
the shell script and keeps only the JSON parsing and filtering
logic.
